Shift the spectrum function to base class, away from GMRES.
